backend/requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
Django==4.1.2
django-appconf==1.0.5
django-cors-headers==3.13.0
django-cryptography==1.1
django-encrypted-files==0.0.10
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.4
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
iniconfig==1.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
psycopg2-binary==2.9.4
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.4
urllib3==1.26.13
wincertstore==0.2
